+++ frontend/src/components/organisms/index.ts
@@ -3,3 +3,4 @@ export { default as Nav } from "./Nav";
 export { default as SidebarContainer } from "./SidebarContainer";
 export { default as SearchFilters } from "./SearchFilters";
 export { default as InstanceTable } from "./InstanceTable";
+export { default as FederationTab } from "./FederationTab";